+++ VERMAXION/Configuration.cs
@@ -22,7 +22,7 @@ public class Configuration : IPluginConfiguration
     public bool KrangleEnabled { get; set; } = false;
     public bool AutoRestoreRetainerCheckingAfterWork { get; set; } = true;
     public bool EnableCharacterSelectStallRecovery { get; set; } = true;
-    public int RefillListingsDelayAdjustmentMs { get; set; } = 0;
+    public int RefillListingsActionDelayMs { get; set; } = 250;
     public float LeftPanelWidth { get; set; } = 240f;
     public CharacterListSortMode CharacterListSortMode { get; set; } = CharacterListSortMode.Name;
     public List<string> PostProcessTaskOrder { get; set; } = VERMAXION.PostProcessTaskOrder.DefaultOrder.ToList();